Bug fixes, preparing for rendering
I tried to be clever with the drawing, but as far as I can tell that isn't going to work. I'll have to do things the way the Lua version did them.

Also, there's a problem right now with Font.c. If you draw several different lines of text in debug mode, it will crash with a sokol error about not updating the same image multiple times per frame. It works fine in release mode, because validation is disabled for sokol. Maybe move to sokol_fontstash for now to fix this?