Experimenting with simpler visuals
Just messing with ways to make the structured rendering be less verbose than normal text. This implementation is far from perfect but it shows how things could look without lots of boxes and padding. The changes here a pretty hacky so there are still bugs to fix with the spacing. A better version could also show backgrounds or highlights for nearby blocks at the same AST depth to make navigation easier.